Catch invalid saveEntityRecord response types and throw error
Throwing here causes a rejected Promise to be returned from `createEntity`. This is then caught and handled inside `LinkControl` in order to centralise error handling.

This area will need additional safeguards adding to catch response types which are errors or which don’t confirm to the expected format of `entity`.